refactor(configure): change configure checker location
